packages/watchmen-ai-copilot/pyproject.toml
[tool.poetry]
name = "watchmen-ai-copilot"
version = "0.1.0"
description = ""
authors = ["luke0623 <luke0623@outlook.com>"]

packages = [
    { include = "watchmen_ai", from = "src" }
]

[tool.poetry.dependencies]
python = "~3.12"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
python-multipart = "^0.0.9"
watchmen-lineage = { path = "../watchmen-lineage", develop = true }
watchmen-model = { path = "../watchmen-model", develop = true }
watchmen-meta = { path = "../watchmen-meta", develop = true }
watchmen-data-surface = { path = "../watchmen-data-surface", develop = true }
watchmen-inquiry-surface = { path = "../watchmen-inquiry-surface", develop = true }
watchmen-indicator-surface = { path = "../watchmen-indicator-surface", develop = true }
watchmen-storage-mysql = { path = "../watchmen-storage-mysql", develop = true }
watchmen-storage-oracle = { path = "../watchmen-storage-oracle", develop = true, optional = true }
watchmen-storage-mssql = { path = "../watchmen-storage-mssql", develop = true, optional = true }
watchmen-storage-postgresql = { path = "../watchmen-storage-postgresql", develop = true, optional = true }
kafka-python = { version = "^2.0.2", optional = true }
aiokafka = { version = "^0.11.0", optional = true }
aio-pika = { version = "^9.4.2", optional = true }
starlette-prometheus = { version = "^0.9.0", optional = true }
requests = { version = "^2.32.3", optional = true }
python3-saml = { version = "^1.16.0", optional = true }
cryptography = { version = "^42.0.8", optional = true }
uvloop = "^0.19.0"
pydantic-settings = "^2.7.1"
python-json-logger = "2.0.7"
sqlparse = "^0.5.3"
multidict = "^6.1.0"
yarl = "^1.18.3"
botbuilder = "^0.0.1"
botbuilder-core = "^4.16.2"
botbuilder-integration-aiohttp = "^4.16.2"
botbuilder-dialogs = "^4.16.2"
adaptivecardbuilder = "^0.0.9"
kuzu = "^0.7.1"
langchain-community = "^0.3.14"
plotly = "^5.24.1"
retry = "^0.9.2"
icecream = "^2.1.3"
tabulate = "^0.9.0"
humanize = "^4.12.2"
scipy = "^1.15.2"
scikit-learn = "^1.6.1"
langgraph = "0.2.52"
langchain-mcp-tools = "^0.2.4"
langchain = "^0.3.25"
langchain-openai = "^0.3.21"
blinker = "^1.9.0"
lancedb = "^0.14.0"
sentence-transformers = "^3.3.1"
pyarrow = "^18.1.0"
numpy = "^2.2.1"
openai = "^1.0.0"

[tool.poetry.group.dev.dependencies]


[tool.poetry.extras]
trino = ["watchmen-inquiry-trino"]
mysql = ["watchmen-storage-mysql", "cryptography"]
oracle = ["watchmen-storage-oracle"]
mongodb = ["watchmen-storage-mongodb"]
mssql = ["watchmen-storage-mssql"]
postgresql = ["watchmen-storage-postgresql"]
oss = ["watchmen-storage-oss"]
s3 = ["watchmen-storage-s3"]
adls = ["watchmen-storage-adls"]
collector = ["watchmen-collector-surface"]
kafka = ["kafka-python", "aiokafka"]
rabbit = ["aio-pika"]
prometheus = ["starlette-prometheus"]
standard_ext_writer = ["requests"]
es_ext_writer = []
sso = ["python3-saml", "cryptography"]
boto3 = ["boto3"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
